Use a work queue per RealmCoordinator rather than a single global queue.
This removes the possibility for a subscription operation on one Realm
to be blocked by another subscription operation on a Realm that's in a
write transaction.
