+++ src/impl/realm_coordinator.hpp
@@ -38,6 +38,10 @@ class CollectionNotifier;
 class ExternalCommitHelper;
 class WeakRealmNotifier;
 
+namespace partial_sync {
+class WorkQueue;
+}
+
 // RealmCoordinator manages the weak cache of Realm instances and communication
 // between per-thread Realm instances for a given file
 class RealmCoordinator : public std::enable_shared_from_this<RealmCoordinator> {
@@ -137,6 +141,11 @@ public:
     template<typename Pred>
     std::unique_lock<std::mutex> wait_for_notifiers(Pred&& wait_predicate);
 
+#if REALM_ENABLE_SYNC
+    // A work queue that can be used to perform background work related to partial sync.
+    partial_sync::WorkQueue& partial_sync_work_queue();
+#endif
+
 private:
     Realm::Config m_config;
 
@@ -170,7 +179,10 @@ private:
     std::unique_ptr<_impl::ExternalCommitHelper> m_notifier;
     std::function<void(VersionID, VersionID)> m_transaction_callback;
 
+#if REALM_ENABLE_SYNC
     std::shared_ptr<SyncSession> m_sync_session;
+    std::unique_ptr<partial_sync::WorkQueue> m_partial_sync_work_queue;
+#endif
 
     // must be called with m_notifier_mutex locked
     void pin_version(VersionID version);
